Refactor monitor runner logic for separation of concerns and better testability.
