+++ alerting/src/main/kotlin/org/opensearch/alerting/transport/TransportExecuteMonitorAction.kt
@@ -15,7 +15,7 @@ import org.opensearch.action.get.GetRequest
 import org.opensearch.action.get.GetResponse
 import org.opensearch.action.support.ActionFilters
 import org.opensearch.action.support.HandledTransportAction
-import org.opensearch.alerting.MonitorRunner
+import org.opensearch.alerting.MonitorRunnerService
 import org.opensearch.alerting.action.ExecuteMonitorAction
 import org.opensearch.alerting.action.ExecuteMonitorRequest
 import org.opensearch.alerting.action.ExecuteMonitorResponse
@@ -39,11 +39,11 @@ import java.time.Instant
 private val log = LogManager.getLogger(TransportGetMonitorAction::class.java)
 
 class TransportExecuteMonitorAction @Inject constructor(
-    transportService: TransportService,
-    private val client: Client,
-    private val runner: MonitorRunner,
-    actionFilters: ActionFilters,
-    val xContentRegistry: NamedXContentRegistry
+        transportService: TransportService,
+        private val client: Client,
+        private val runner: MonitorRunnerService,
+        actionFilters: ActionFilters,
+        val xContentRegistry: NamedXContentRegistry
 ) : HandledTransportAction<ExecuteMonitorRequest, ExecuteMonitorResponse> (
     ExecuteMonitorAction.NAME, transportService, actionFilters, ::ExecuteMonitorRequest
 ) {
@@ -63,11 +63,7 @@ class TransportExecuteMonitorAction @Inject constructor(
                     val (periodStart, periodEnd) =
                         monitor.schedule.getPeriodEndingAt(Instant.ofEpochMilli(execMonitorRequest.requestEnd.millis))
                     try {
-                        val monitorRunResult = if (monitor.isBucketLevelMonitor()) {
-                            runner.runBucketLevelMonitor(monitor, periodStart, periodEnd, execMonitorRequest.dryrun)
-                        } else {
-                            runner.runQueryLevelMonitor(monitor, periodStart, periodEnd, execMonitorRequest.dryrun)
-                        }
+                        val monitorRunResult = runner.runJob(monitor, periodStart, periodEnd, execMonitorRequest.dryrun)
                         withContext(Dispatchers.IO) {
                             actionListener.onResponse(ExecuteMonitorResponse(monitorRunResult))
                         }
